refactor: 폴더 구조 재정리 및 컴포넌트 네이밍
