+++ src/pages/MainPage.tsx
@@ -1,14 +1,14 @@
-import { AIGift } from '@/components/feature/Home/AIGift';
-import { GiftRanking } from '@/components/feature/Home/GiftRanking';
-import { SelectFriend } from '@/components/feature/Home/SelectFriend';
-import { ThemeCategory } from '@/components/feature/Home/ThemeCategory';
+import { AIRecommendedGiftBanner } from '@/components/Home/AIRecommendedGiftBanner';
+import { GiftRanking } from '@/components/Home/GiftRanking';
+import { SelectFriend } from '@/components/Home/SelectFriend';
+import { ThemeCategory } from '@/components/Home/ThemeCategory';
 
 export default function MainPage() {
   return (
     <>
       <SelectFriend />
       <ThemeCategory />
-      <AIGift />
+      <AIRecommendedGiftBanner />
       <GiftRanking />
     </>
   );
