+++ src/pages/ThemePage.tsx
@@ -1,7 +1,7 @@
 import { useParams } from 'react-router-dom';
 
-import { HeaderSection } from '@/components/feature/Theme/HeaderSection';
-import { ItemList } from '@/components/feature/Theme/ItmeList';
+import { HeaderSection } from '@/components/Theme/HeaderSection';
+import { ItemList } from '@/components/Theme/ItmeList';
 
 export default function ThemePage() {
   const { themeKey = '' } = useParams<{ themeKey: string }>();
